requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & Agent Framework
langchain>=0.1.0
langgraph>=0.0.32
openai>=1.0.0

# HTTP Client
httpx>=0.25.0

# Vector Database
qdrant-client>=1.6.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Document Processing
unstructured[all-docs]>=0.10.0
pypdf2>=3.0.0
pdfplumber>=0.9.0

# Utils
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
